+++ pkg/executor/middleware/ratelimit/compute_units.go
@@ -8,12 +8,12 @@ import (
 
 	"go.uber.org/zap"
 
+	"github.com/superblocksteam/agent/internal/flags"
 	imetrics "github.com/superblocksteam/agent/internal/metrics"
 	contextPkg "github.com/superblocksteam/agent/pkg/context"
 	sberrors "github.com/superblocksteam/agent/pkg/errors"
 	"github.com/superblocksteam/agent/pkg/executor/middleware"
 	"github.com/superblocksteam/agent/pkg/executor/options"
-	"github.com/superblocksteam/agent/pkg/flags"
 	"github.com/superblocksteam/agent/pkg/metrics"
 	"github.com/superblocksteam/agent/pkg/observability"
 	"github.com/superblocksteam/agent/pkg/store"